Refactor/reorg
