+++ src/Text/Text.tsx
@@ -1,7 +1,7 @@
 import { createElement, ElementType, HTMLAttributes, forwardRef } from 'react';
 import cx from 'clsx';
-import { Color } from '../color';
-import { Opacity } from '../scale';
+import { Color } from '../tokens/color';
+import { Opacity } from '../tokens/scale';
 import {
   FontFamily,
   FontSize,
@@ -27,7 +27,7 @@ import {
   createPrefixValueResolver,
   resolveValue,
 } from '../utils';
-import { resolveTextOverflow } from './utils';
+import { resolveTextOverflow } from '../utils/text';
 
 export type TextProps = HTMLAttributes<HTMLElement> & {
   element?: ElementType;
