+++ src/utils/index.ts
@@ -1,2 +1,4 @@
 export * from './responsive';
 export * from './resolvers';
+export * from './margin';
+export * from './padding';
